+++ anyio/_backends/_asyncio.py
@@ -187,8 +187,8 @@ class CancelScope:
             # Cancel the task directly, but only if it's blocked and isn't within a shielded scope
             cancel_scope = _task_states[task].cancel_scope
             if cancel_scope is self:
-                # Only deliver the cancellation if the task is already running
-                if task._coro.cr_await is not None:
+                # Only deliver the cancellation if the task is already running (but not this task!)
+                if not task._coro.cr_running and task._coro.cr_await is not None:
                     task.cancel()
             elif not cancel_scope._shielded_to(self):
                 await cancel_scope._cancel()